requirements.txt
gunicorn
flask
newspaper3k
lxml_html_clean
nltk
textblob
validators
requests
torch
torchaudio
transformers
datasets
evaluate
accelerate
peft
scikit-learn
bert-score
sentencepiece
protobuf
ruff
PyPDF2
easyocr
librosa
moviepy
matplotlib
